Refactor connect organization functionality into another component
